Move some modules out of game
