+++ src/game/mod.rs
@@ -3,17 +3,9 @@
 use bevy::prelude::*;
 
 mod animation;
-pub mod assets;
-pub mod audio;
 mod movement;
 pub mod spawn;
 
 pub(super) fn plugin(app: &mut App) {
-    app.add_plugins((
-        animation::plugin,
-        audio::plugin,
-        assets::plugin,
-        movement::plugin,
-        spawn::plugin,
-    ));
+    app.add_plugins((animation::plugin, movement::plugin, spawn::plugin));
 }
