+++ src/screens/credits.rs
@@ -4,8 +4,8 @@ use bevy::prelude::*;
 
 use super::Screen;
 use crate::{
-    game::{assets::SoundtrackHandles, audio::soundtrack::SoundtrackCommands as _},
     ui::prelude::*,
+    assets::SoundtrackHandles, audio::soundtrack::SoundtrackCommands as _,
 };
 
 pub(super) fn plugin(app: &mut App) {
