+++ src/screens/playing.rs
@@ -3,9 +3,8 @@
 use bevy::{input::common_conditions::input_just_pressed, prelude::*};
 
 use super::Screen;
-use crate::game::{
-    assets::SoundtrackHandles, audio::soundtrack::SoundtrackCommands as _, spawn::level::SpawnLevel,
-};
+use crate::game::spawn::level::SpawnLevel;
+use crate::{assets::SoundtrackHandles, audio::soundtrack::SoundtrackCommands as _};
 
 pub(super) fn plugin(app: &mut App) {
     app.add_systems(OnEnter(Screen::Playing), spawn_level);
